examples/zeromq/requirements.txt
pyzmq >= 26.0